requirements.txt
delta-spark==4.0.0
importlib_metadata==8.7.0
iniconfig==2.3.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
py4j==0.10.9.9
Pygments==2.19.2
pyspark==4.0.1
pytest==9.0.1
ruff==0.14.4
typing_extensions==4.15.0
zipp==3.23.0